split up and finish implementation and tests
