+++ dev-tools/src/docker.test.ts
@@ -1,15 +1,8 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { vol } from "memfs";
-import {
-  // generateDockerfile,
-  getMonorepoPackages,
-  buildWorkspaceDependencyGraph,
-  findPackagesWithDockerfileTemplates,
-  getAllPackageWorkspaceDependencies,
-  buildMonorepoContext,
-  generateDockerfiles,
-} from "./docker.ts";
+import { generateDockerfiles } from "./docker.ts";
 import { monorepoPackageMock } from "./monorepo.mock.ts";
+import { buildMonorepoContext } from "./workspace.ts";
 vi.mock("node:fs");
 vi.mock("node:fs/promises");
 
@@ -21,73 +14,19 @@ afterEach(() => {
   vol.reset();
 });
 
-describe("getMonorepoPackageJsons", () => {
-  it("should return all workspace packages", () => {
-    const { monorepoPackageJsons } = getMonorepoPackages("/app");
-    expect(monorepoPackageJsons).toBeDefined();
-    expect(monorepoPackageJsons["@foo/foo"]).toBeDefined();
-    expect(monorepoPackageJsons["@foo/foo"].workspaces).toBeDefined();
-
-    // it should gather specific packages listed in workspaces
-    expect(monorepoPackageJsons["@foo/main-db"]).toBeDefined();
-    expect(
-      monorepoPackageJsons["@foo/main-db"].dependencies?.["third-party-lib"],
-    ).toBeDefined();
-
-    // it should gather packages in sub-folders
-    expect(monorepoPackageJsons["@foo/api-service"]).toBeDefined();
-    expect(monorepoPackageJsons["@saflib/node-express"]).toBeDefined();
-  });
-});
-
-describe("buildWorkspaceDependencyGraph", () => {
-  it("should return the correct dependency graph", () => {
-    const { monorepoPackageJsons } = getMonorepoPackages("/app");
-    const dependencyGraph = buildWorkspaceDependencyGraph(monorepoPackageJsons);
-    expect(dependencyGraph).toBeDefined();
-    expect(dependencyGraph["@foo/auth-web-client"]).toStrictEqual([
-      "@saflib/vue-spa",
-      "@saflib/auth-vue",
-    ]);
-  });
-});
-
-describe("findPackagesWithDockerfileTemplates", () => {
-  it("should return the correct packages", () => {
-    const { monorepoPackageDirectories } = getMonorepoPackages("/app");
-    const packages = findPackagesWithDockerfileTemplates(
-      monorepoPackageDirectories,
-    );
-    expect(packages).toStrictEqual([
-      "@foo/auth-web-client",
-      "@foo/www-web-client",
-      "@foo/api-service",
-      "@foo/auth-service",
-    ]);
-  });
-});
-
-describe("getAllPackageWorkspaceDependencies", () => {
-  it("should return the correct dependencies", () => {
-    const context = buildMonorepoContext("/app");
-    const dependencies = getAllPackageWorkspaceDependencies(
-      "@foo/auth-web-client",
-      context,
-    );
-    expect(dependencies).toStrictEqual(
-      new Set([
-        "@saflib/vue-spa",
-        "@saflib/auth-vue",
-        "@saflib/auth-spec",
-        "@saflib/openapi-specs",
-      ]),
-    );
-  });
-});
-
 describe("generateDockerfiles", () => {
   it("should generate the correct dockerfiles", () => {
     const context = buildMonorepoContext("/app");
     generateDockerfiles(context);
+    const dockerfile = vol.readFileSync(
+      "/app/clients/web-auth/Dockerfile",
+      "utf-8",
+    );
+    expect(dockerfile).toContain(
+      "COPY --parents ./package.json ./package-lock.json ./saflib/vue-spa/package.json ./saflib/auth-vue/package.json ./saflib/auth-spec/package.json ./saflib/openapi-specs/package.json ./clients/web-auth/package.json ./",
+    );
+    expect(dockerfile).toContain(
+      "COPY --parents ./saflib/vue-spa ./saflib/auth-vue ./saflib/auth-spec ./saflib/openapi-specs ./clients/web-auth ./",
+    );
   });
 });
